[tests] move user frontend test bases to their usages
This shall aid in converting the setup to pytest fixtures.

Preps for #580-related action.
