+++ tests/frontend/user/test_logs.py
@@ -1,9 +1,35 @@
+from flask import url_for
+
 from hades_logs import HadesLogs
 from tests.factories import UserFactory, RoomLogEntryFactory, \
     UserLogEntryFactory
-from . import UserLogTestBase
-from ..legacy_base import InvalidateHadesLogsMixin
+from ..legacy_base import InvalidateHadesLogsMixin, FrontendWithAdminTestBase
 from ...hades_logs import get_hades_logs_config
+
+
+class UserLogTestBase(FrontendWithAdminTestBase):
+    def get_logs(self, user_id=None, **kw):
+        """Request the logs, assert validity, and return the response.
+
+        By default, the logs are fetched for the user logging in.
+
+        The following assertions are made:
+          * The response code is 200
+          * The response content_type contains ``"json"``
+          * The response's JSON contains an ``"items"`` key
+
+        :returns: ``response.json['items']``
+        """
+        if user_id is None:
+            user_id = self.user_id
+        log_endpoint = url_for('user.user_show_logs_json',
+                               user_id=user_id,
+                               **kw)
+        response = self.assert_response_code(log_endpoint, code=200)
+        assert "json" in response.content_type.lower()
+        json = response.json
+        assert json.get('items') is not None
+        return json['items']
 
 
 class AppWithoutHadesLogsTestCase(InvalidateHadesLogsMixin, UserLogTestBase):
